Move the Keywords after the Mailing list header
At the moment, MELPA's archive-contents contains the mailing list
information in the keywords because they're using an older version of
Emacs for building archive-contents.

I have verified that this does not break the Package-Requires header.

This bug has been fixed in Emacs in
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=13207.

The issue for MELPA is https://github.com/melpa/melpa/issues/4087.