Use ‘evil-surround’ as :group for customizable variables
Before this change, there were two groups for this package in the
customize interface. The customizable variables show up in the group
named ‘surround’, while the mode itself shows up in the group named
‘evil-surround’. This is confusing and unnecessary.

To address this, put everything inside the ‘evil-surround’ group,
which is the same approach used by many other evil-mode extension
packages.