Fix repeat when pair is determined by function
Apparently when read-from-minibuffer is called in a function to read tags, the
information in this-command-keys is discarded. Introduce
evil-surround-read-from-minibuffer to record this information as well as the
input in the minibuffer.

There is probably a more elegant solution to recording repeat info in
evil-surround out there, but this will work for now.

Fixes #125