Add a test for surrounding with a tag in a macro
Not using `evil-test-buffer`, as it doesn't seem to play well with
this kind of testing a kbd macro.

Also slightly different from the already-present tests for surrounding
with tags, because it does additional things after wrapping -- without
the included change, this surrounds with a `<div>` tag, but also
inserts the following keys as well, instead of treating them as
commands.