Limit global activation to text-, prog- and comint-mode
Activating evil-surround in special-mode (i.e. read-only) buffers does not make
sense. Furthermore, the bindings can overwrite bindings in modes like
magit-diff-mode (see
https://github.com/syl20bnr/spacemacs/issues/15448#issuecomment-1090810486).

This commit limits the 'globally activated modes' to (modes derived from) text-,
prog- and comint-mode via the `define-globalized-minor-mode` its
[:predicate](https://www.gnu.org/software/emacs/manual/html_node/elisp/Defining-Minor-Modes.html#index-define_002dglobalized_002dminor_002dmode]
keyword.

As this should be the default behavior, it is not essential to change/update the
documentation.